// apps/web/src/proto/Permission.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Permission = "server_manage" | "channel_create" | "channel_edit" | "channel_delete" | "channel_join_locked" | "client_move" | "client_kick" | "token_create" | "client_ban" | "group_manage" | "message_manage" | "invite_create" | "file_upload" | "member_remove" | "member_nickname";